requirements.txt
astrapy
python-dotenv
openai
pandas
streamlit